requirements.txt
opencv-python
pandas
timm
pytorch_pretrained_vit
scipy